Build for TARGET_ARCH=fusion_f1 via reference kernel fallbacks.
This change adds reference fallbacks to the optimized xtensa kernels for
the case when TARGET_ARCH is anything other than hifimini.

This sets the stage for a baseline from which we can incrementally
optimize for architectures other than hifimini.

The goal is to have a starting point where all the unit tests pass for
`TARGET_ARCH=hifimini` (which will use the optimized implementations) or
any other `TARGET_ARCH` (with reference fallback).

Tested for `TARGET_ARCH=fusion_f1` with:

```
make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=fusion_f1 XTENSA_CORE=Google_F1 test
```

With the following profiling results:

```
make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=fusion_f1 XTENSA_CORE=Google_F1 test_keyword_benchmark

InitializeKeywordRunner() took 239061 ticks (239 ms)
KeywordRunNIerations(1) took 168564 ticks (168 ms)
KeywordRunNIerations(10) took 1685111 ticks (1685 ms)
```

```
make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=fusion_f1 XTENSA_CORE=Google_F1 keyword_benchmark BUILD_TYPE=release
xt-size tensorflow/lite/micro/tools/make/gen/xtensa_fusion_f1/bin/keyword_benchmark

   text	   data	    bss	    dec	    hex	filename
  48256	  40132	  24952	 113340	  1babc	tensorflow/lite/micro/tools/make/gen/xtensa_fusion_f1/bin/keyword_benchmark
```

After this change, we can:
 * add a continuous build for Hifi4
 * add optimizations for Hifi4 on a per kernelbasis and keep profiling
   the impact of these optimizations on the keyword benchmark cycles and
   binary size.

Also tested that `TARGET_ARCH=hifimini` is unaffected:

```
make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=hifimini XTENSA_CORE=mini1m1m_RG test_keyword_benchmark

InitializeKeywordRunner() took 1392788 ticks (1392 ms)
KeywordRunNIerations(1) took 89195 ticks (89 ms)
KeywordRunNIerations(10) took 891509 ticks (891 ms)
```

```
make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=hifimini XTENSA_CORE=mini1m1m_RG keyword_benchmark BUILD_TYPE=release
xt-size tensorflow/lite/micro/tools/make/gen/xtensa_hifimini/bin/keyword_benchmark
   text	   data	    bss	    dec	    hex	filename
  46080	  40204	  24952	 111236	  1b284	tensorflow/lite/micro/tools/make/gen/xtensa_hifimini/bin/keyword_benchmark
```
